+++ src/plugins/pouchdb/index.ts
@@ -2,3 +2,6 @@ export * from './pouch-db';
 export * from './rx-storage-pouchdb';
 export * from './adapter-check';
 export * from './custom-events-plugin';
+export * from './pouchdb-helper';
+export * from './rx-storage-instance-pouch';
+export * from './rx-storage-key-object-instance-pouch';